extract plp data hook
