+++ packages/extension-chakra-storefront/src/pages/product-list/index.jsx
@@ -5,12 +5,7 @@
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
-import React, {useEffect, useState} from 'react'
-import {useLocation, useParams} from 'react-router-dom'
-import {keepPreviousData} from '@tanstack/react-query'
-import {useCategory, useProductSearch} from '@salesforce/commerce-sdk-react'
-
-// Components
+import React from 'react'
 import {Box, Grid, Stack} from '@chakra-ui/react'
 
 // Project Components
@@ -25,22 +20,9 @@ import ProductListGrid from './partials/product-list-grid'
 import ProductListPagination from './partials/product-list-pagination'
 
 // Hooks
-import {usePageUrls, useSortUrls, useSearchParams, useExtensionConfig} from '../../hooks'
-import useEinstein from '../../hooks/use-einstein'
-import useActiveData from '../../hooks/use-active-data'
-import useDataCloud from '../../hooks/use-datacloud'
+import {useExtensionConfig} from '../../hooks'
 import {useProductListWishlist} from './hooks/use-product-list-wishlist'
-
-// Others
-import {HTTPNotFound, HTTPError} from '@salesforce/pwa-kit-react-sdk/ssr/universal/errors'
-import logger from '../../utils/logger-instance'
-
-// Constants
-import useNavigation from '../../hooks/use-navigation'
-
-// NOTE: You can ignore certain refinements on a template level by updating the below
-// list of ignored refinements.
-const REFINEMENT_DISALLOW_LIST = ['c_isNew']
+import {useProductListData} from './hooks/use-product-list-data'
 
 /*
  * This is a simple product listing page. It displays a paginated list
@@ -48,213 +30,32 @@ const REFINEMENT_DISALLOW_LIST = ['c_isNew']
  * allowable filters and sort refinements.
  */
 const ProductList = () => {
-    const navigate = useNavigation()
-    const params = useParams()
-    const location = useLocation()
-    const einstein = useEinstein()
-    const dataCloud = useDataCloud()
-    const activeData = useActiveData()
-    const [searchParams, {stringify: stringifySearchParams}] = useSearchParams()
     const {
         pages: {ProductList: productListConfig}
     } = useExtensionConfig()
-    /**************** Page State ****************/
-    const [filtersLoading, setFiltersLoading] = useState(false)
     const {addItem, removeItem, isItemInWishlist} = useProductListWishlist()
 
-    const urlParams = new URLSearchParams(location.search)
-    let searchQuery = urlParams.get('q')
-    const isSearch = !!searchQuery
-
-    if (params.categoryId) {
-        searchParams._refine.push(`cgid=${params.categoryId}`)
-    }
-
-    /**************** Mutation Actions ****************/
-
-    /**************** Query Actions ****************/
-    // _refine is an invalid param for useProductSearch, we don't want to pass it to API call
-    const {_refine, ...restOfParams} = searchParams
-
     const {
-        isLoading,
+        basePath,
+        category,
+        filtersLoading,
+        handleProductClick,
         isFetched,
+        isLoading,
         isRefetching,
-        data: productSearchResult
-    } = useProductSearch(
-        {
-            parameters: {
-                ...restOfParams,
-                perPricebook: true,
-                allVariationProperties: true,
-                allImages: true,
-                expand: [
-                    'promotions',
-                    'variations',
-                    'prices',
-                    'images',
-                    'page_meta_tags',
-                    'custom_properties'
-                ],
-                refine: _refine
-            }
-        },
-        {
-            placeholderData: keepPreviousData
-        }
-    )
-
-    const {error, data: category} = useCategory(
-        {
-            parameters: {
-                id: params.categoryId
-            }
-        },
-        {
-            enabled: !isSearch && !!params.categoryId
-        }
-    )
-
-    // Apply disallow list to refinements.
-    if (productSearchResult?.refinements) {
-        productSearchResult.refinements = productSearchResult.refinements.filter(
-            ({attributeId}) => !REFINEMENT_DISALLOW_LIST.includes(attributeId)
-        )
-    }
-
-    /**************** Error Handling ****************/
-    const errorStatus = error?.response?.status
-    switch (errorStatus) {
-        case undefined:
-            // No Error.
-            break
-        case 404:
-            throw new HTTPNotFound('Category Not Found.')
-        default:
-            throw new HTTPError(errorStatus, `HTTP Error ${errorStatus} occurred.`)
-    }
-
-    // Reset scroll position when `isRefetching` becomes `true`.
-    useEffect(() => {
-        isRefetching && window.scrollTo(0, 0)
-        setFiltersLoading(isRefetching)
-    }, [isRefetching])
-
-    /**************** Render Variables ****************/
-    const basePath = `${location.pathname}${location.search}`
-    const showNoResults = !isLoading && productSearchResult && !productSearchResult?.hits
-    const {total, sortingOptions} = productSearchResult || {}
-
-    // Get urls to be used for pagination, page size changes, and sorting.
-    const pageUrls = usePageUrls({total})
-    const sortUrls = useSortUrls({options: sortingOptions})
-
-    /**************** Action Handlers ****************/
-    // Toggles filter on and off
-    const toggleFilter = (value, attributeId, selected, allowMultiple = true) => {
-        const searchParamsCopy = {...searchParams}
-
-        // Remove the `offset` search param if present.
-        delete searchParamsCopy.offset
-
-        // If we aren't allowing for multiple selections, simply clear any value set for the
-        // attribute, and apply a new one if required.
-        if (!allowMultiple) {
-            const previousValue = searchParamsCopy.refine[attributeId]
-            delete searchParamsCopy.refine[attributeId]
-
-            // Note the loose comparison, for "string != number" checks.
-            if (!selected && value.value != previousValue) {
-                searchParamsCopy.refine[attributeId] = value.value
-            }
-        } else {
-            // Get the attibute value as an array.
-            let attributeValue = searchParamsCopy.refine[attributeId] || []
-
-            // Ensure that the value is still converted into an array if it's a `string` or `number`.
-            if (typeof attributeValue === 'string') {
-                attributeValue = attributeValue.split('|')
-            } else if (typeof attributeValue === 'number') {
-                attributeValue = [attributeValue]
-            }
-
-            // Either set the value, or filter the value out.
-            if (!selected) {
-                attributeValue.push(value.value)
-            } else {
-                // Note the loose comparison, for "string != number" checks.
-                attributeValue = attributeValue?.filter((v) => v != value.value)
-            }
-
-            // Update the attribute value in the new search params.
-            searchParamsCopy.refine[attributeId] = attributeValue
-
-            // If the update value is an empty array, remove the current attribute key.
-            if (searchParamsCopy.refine[attributeId].length === 0) {
-                delete searchParamsCopy.refine[attributeId]
-            }
-        }
-
-        if (isSearch) {
-            navigate(`/search?${stringifySearchParams(searchParamsCopy)}`)
-        } else {
-            navigate(`/category/${params.categoryId}?${stringifySearchParams(searchParamsCopy)}`)
-        }
-    }
-
-    // Clears all filters
-    const resetFilters = () => {
-        const newSearchParams = {
-            ...searchParams,
-            refine: []
-        }
-        const newPath = isSearch
-            ? `/search?${stringifySearchParams(newSearchParams)}`
-            : `/category/${params.categoryId}?${stringifySearchParams(newSearchParams)}`
-
-        navigate(newPath)
-    }
-
-    /**************** Einstein ****************/
-    useEffect(() => {
-        if (productSearchResult) {
-            if (isSearch) {
-                try {
-                    einstein.sendViewSearch(searchQuery, productSearchResult)
-                } catch (err) {
-                    logger.error('Einstein sendViewSearch error', {
-                        namespace: 'ProductList.useEffect',
-                        additionalProperties: {error: err, searchQuery}
-                    })
-                }
-                dataCloud.sendViewSearchResults(searchParams, productSearchResult)
-                activeData.sendViewSearch(searchParams, productSearchResult)
-            } else {
-                try {
-                    einstein.sendViewCategory(category, productSearchResult)
-                } catch (err) {
-                    logger.error('Einstein sendViewCategory error', {
-                        namespace: 'ProductList.useEffect',
-                        additionalProperties: {error: err, category}
-                    })
-                }
-                dataCloud.sendViewCategory(searchParams, category, productSearchResult)
-                activeData.sendViewCategory(searchParams, category, productSearchResult)
-            }
-        }
-    }, [productSearchResult])
+        pageUrls,
+        productSearchResult,
+        resetFilters,
+        searchQuery,
+        searchParams,
+        showNoResults,
+        sortUrls,
+        toggleFilter
+    } = useProductListData()
 
     const handleFavouriteToggle = (product, isFavourite) => {
         const action = isFavourite ? addItem : removeItem
         action(product)
-    }
-
-    const handleProductClick = (product) => {
-        if (searchQuery) {
-            einstein.sendClickSearch(searchQuery, product)
-        } else if (category) {
-            einstein.sendClickCategory(category, product)
-        }
     }
 
     return (
